Extended the terms of the license on line 15
This is not a license to remote control the Human author of this LICENSE file using any form of technology known or unknown (e.g. EMF, RF, THz waves, etc...). No `Human being`, `Earthling`, or `i§i (integral §entient intelligence)` is to be recognized as a `software` by autonomous agents attempting to use this LICENSE file.